pyproject.toml
[project]
name = "nonebot-plugin-myhoyotool"
dynamic = ["version"]
description = "one of PDM package"
authors = [
    { name = "scriptsmay", email = "scriptsmay@163.com" },
]
dependencies = [
    "nonebot2>=2.1.1",
    "pydantic>=2.9.0",
    "nonebot-adapter-onebot>=2.3.1",
    "nonebot-plugin-htmlrender>=0.2.2",
    "nonebot-plugin-apscheduler>=0.3.0",
    "nonebot-plugin-send-anything-anywhere>=0.4.0",
    "aiohttp >=3.9.1",
    "pydantic-settings>=2.6.0",
    "httpx>=0.27.2",
    "Pillow>=11.0.0",
]
requires-python = ">=3.12"
readme = "README.md"
license = { text = "MIT" }
keywords = ["star rail","mihoyo", "genshin", "nonebot", "nonebot2", "bot", "qq", "telegram"]
classifiers = [
    "Natural Language :: Chinese (Simplified)",
    "Topic :: Communications",
    "Topic :: Communications :: Chat",
    "Topic :: Games/Entertainment",
    "Topic :: Internet",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content"
]

[project.urls]
homepage = "https://github.com/scriptsmay/nonebot-plugin-myhoyotool"

[project.optional-dependencies]
menu = ["nonebot-plugin-PicMenu>=0.2"]

[tool.pdm]
distribution = true
[tool.pdm.version]
source = "file"
path = "nonebot_plugin_myhoyotool/__init__.py"

[tool.pdm.build]
includes = []

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"
